requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
python-decouple
passlib
bcrypt
requests
pika # For RabbitMQ
python-dotenv
PyJWT
alembic
requests
python-jose
passlib[bcrypt]
email-validator